Refactoring: split track.py file
Separate Segment and Track related code.
